Add RootLayout component for application structure
- Introduced a new RootLayout component in app/layout.tsx to serve as the main layout for the application.
- The component includes metadata for the application and integrates a ThemeProvider for consistent theming.
- Simplified the layout structure in the (auth), (authenticated), and (public) directories by removing unnecessary elements and returning only the children.
- This update enhances the overall layout management and prepares the application for further development.
